--- at.jku.se.prse.example.calculator/src/main/java/at/jku/se/prse/example/calculator/operators/NullOperation.java
@@ -1,19 +1,0 @@
-package at.jku.se.prse.example.calculator.operators;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import at.jku.se.prse.example.calculator.factory.ICalculationOperation;
-
-public class NullOperation implements ICalculationOperation {
-
-	private static final Logger LOGGER = LogManager.getLogger(NullOperation.class);
-	
-	@Override
-	public String caclulate(String txt) {
-		LOGGER.warn("This Calculation Operation is not implemented!");
-		return "0";
-
-	}
-
-}
