+++ at.jku.se.prse.team2.logbook/src/main/java/at/jku/se/prse/team2/logbook/business/LogBookEntryFacade.java
@@ -1,0 +1,4 @@
+package at.jku.se.prse.team2.logbook.business;
+
+public class LogBookEntryFacade {
+}
