+++ at.jku.se.prse.team2.logbook/src/main/java/at/jku/se/prse/team2/logbook/factory/ICalculationOperation.java
@@ -1,0 +1,14 @@
+package at.jku.se.prse.team2.logbook.factory;
+
+/**
+ * Calculation Operation interface.<br>
+ * Each Operation provided by the {@link OperationFactory} has to implement this
+ * interface.
+ * 
+ * @author Michael Vierhauser
+ */
+public interface ICalculationOperation {
+
+	String caclulate(String input);
+
+}
